Move to Adapters namespace
